+++ packages/sdk/src/lib/dataProtectorCore/processProtectedData.ts
@@ -1,19 +1,24 @@
 import { ethers } from 'ethers';
 import {
-  DEFAULT_MAX_PRICE,
+  MAX_DESIRED_APP_ORDER_PRICE,
+  MAX_DESIRED_DATA_ORDER_PRICE,
+  MAX_DESIRED_WORKERPOOL_ORDER_PRICE,
   SCONE_TAG,
-  WORKERPOOL_ADDRESS,
 } from '../../config/config.js';
 import {
   WorkflowError,
   processProtectedDataErrorMessage,
   handleIfProtocolError,
 } from '../../utils/errors.js';
-import { fetchOrdersUnderMaxPrice } from '../../utils/fetchOrdersUnderMaxPrice.js';
+import {
+  checkUserVoucher,
+  filterWorkerpoolOrders,
+} from '../../utils/processProtectedData.models.js';
 import { pushRequesterSecret } from '../../utils/pushRequesterSecret.js';
 import {
   addressOrEnsSchema,
   addressSchema,
+  booleanSchema,
   positiveNumberSchema,
   secretsSchema,
   stringSchema,
@@ -22,14 +27,16 @@ import {
   validateOnStatusUpdateCallback,
 } from '../../utils/validators.js';
 import { isERC734 } from '../../utils/whitelist.js';
-import { getResultFromCompletedTask } from './getResultFromCompletedTask.js';
 import {
+  DefaultWorkerpoolConsumer,
+  MatchOptions,
   OnStatusUpdateFn,
   ProcessProtectedDataParams,
   ProcessProtectedDataResponse,
   ProcessProtectedDataStatuses,
 } from '../types/index.js';
-import { IExecConsumer } from '../types/internalTypes.js';
+import { IExecConsumer, VoucherInfo } from '../types/internalTypes.js';
+import { getResultFromCompletedTask } from './getResultFromCompletedTask.js';
 import { getWhitelistContract } from './smartContract/getWhitelistContract.js';
 import { isAddressInWhitelist } from './smartContract/whitelistContract.read.js';
 
@@ -37,17 +44,23 @@ export type ProcessProtectedData = typeof processProtectedData;
 
 export const processProtectedData = async ({
   iexec = throwIfMissing(),
+  defaultWorkerpool,
   protectedData,
   app,
   userWhitelist,
-  maxPrice = DEFAULT_MAX_PRICE,
+  dataMaxPrice = MAX_DESIRED_DATA_ORDER_PRICE,
+  appMaxPrice = MAX_DESIRED_APP_ORDER_PRICE,
+  workerpoolMaxPrice = MAX_DESIRED_WORKERPOOL_ORDER_PRICE,
   path,
   args,
   inputFiles,
   secrets,
   workerpool,
+  useVoucher = false,
+  voucherOwner,
   onStatusUpdate = () => {},
 }: IExecConsumer &
+  DefaultWorkerpoolConsumer &
   ProcessProtectedDataParams): Promise<ProcessProtectedDataResponse> => {
   const vProtectedData = addressOrEnsSchema()
     .required()
@@ -60,9 +73,15 @@ export const processProtectedData = async ({
   const vUserWhitelist = addressSchema()
     .label('userWhitelist')
     .validateSync(userWhitelist);
-  const vMaxPrice = positiveNumberSchema()
-    .label('maxPrice')
-    .validateSync(maxPrice);
+  const vDataMaxPrice = positiveNumberSchema()
+    .label('dataMaxPrice')
+    .validateSync(dataMaxPrice);
+  const vAppMaxPrice = positiveNumberSchema()
+    .label('appMaxPrice')
+    .validateSync(appMaxPrice);
+  const vWorkerpoolMaxPrice = positiveNumberSchema()
+    .label('workerpoolMaxPrice')
+    .validateSync(workerpoolMaxPrice);
   const vPath = stringSchema().label('path').validateSync(path);
   const vInputFiles = urlArraySchema()
     .label('inputFiles')
@@ -70,10 +89,15 @@ export const processProtectedData = async ({
   const vArgs = stringSchema().label('args').validateSync(args);
   const vSecrets = secretsSchema().label('secrets').validateSync(secrets);
   const vWorkerpool = addressOrEnsSchema()
-    .default(WORKERPOOL_ADDRESS) // Default workerpool if none is specified
+    .default(defaultWorkerpool) // Default workerpool if none is specified
     .label('workerpool')
     .validateSync(workerpool);
-
+  const vUseVoucher = booleanSchema()
+    .label('useVoucher')
+    .validateSync(useVoucher);
+  const vVoucherOwner = addressOrEnsSchema()
+    .label('voucherOwner')
+    .validateSync(voucherOwner);
   try {
     const vOnStatusUpdate =
       validateOnStatusUpdateCallback<
@@ -106,63 +130,129 @@ export const processProtectedData = async ({
         requester = vUserWhitelist;
       }
     }
-
-    vOnStatusUpdate({
-      title: 'FETCH_PROTECTED_DATA_ORDERBOOK',
-      isDone: false,
-    });
-    const datasetOrderbook = await iexec.orderbook.fetchDatasetOrderbook(
-      vProtectedData,
-      {
-        app: vApp,
-        workerpool: vWorkerpool,
-        requester,
+    let userVoucher: VoucherInfo | undefined;
+    if (vUseVoucher) {
+      try {
+        userVoucher = await iexec.voucher.showUserVoucher(
+          vVoucherOwner || requester
+        );
+        checkUserVoucher({ userVoucher });
+      } catch (err) {
+        if (err?.message?.startsWith('No Voucher found for address')) {
+          throw new Error(
+            'Oops, it seems your wallet is not associated with any voucher. Check on https://builder.iex.ec/'
+          );
+        }
+        throw err;
       }
-    );
-    vOnStatusUpdate({
-      title: 'FETCH_PROTECTED_DATA_ORDERBOOK',
-      isDone: true,
-    });
+    }
 
     vOnStatusUpdate({
-      title: 'FETCH_APP_ORDERBOOK',
+      title: 'FETCH_ORDERS',
       isDone: false,
     });
-    const appOrderbook = await iexec.orderbook.fetchAppOrderbook(vApp, {
-      dataset: protectedData,
-      requester,
-      minTag: SCONE_TAG,
-      maxTag: SCONE_TAG,
-      workerpool: vWorkerpool,
-    });
+    const [
+      datasetorderForApp,
+      datasetorderForWhitelist,
+      apporder,
+      workerpoolorder,
+    ] = await Promise.all([
+      // Fetch dataset order
+      iexec.orderbook
+        .fetchDatasetOrderbook(vProtectedData, {
+          app: vApp,
+          requester: requester,
+        })
+        .then((datasetOrderbook) => {
+          const desiredPriceDataOrderbook = datasetOrderbook.orders.filter(
+            (order) => order.order.datasetprice <= vDataMaxPrice
+          );
+          return desiredPriceDataOrderbook[0]?.order; // may be undefined
+        }),
+      // Fetch dataset order for whitelist
+      iexec.orderbook
+        .fetchDatasetOrderbook(vProtectedData, {
+          app: vUserWhitelist,
+          requester: requester,
+        })
+        .then((datasetOrderbook) => {
+          const desiredPriceDataOrderbook = datasetOrderbook.orders.filter(
+            (order) => order.order.datasetprice <= vDataMaxPrice
+          );
+          return desiredPriceDataOrderbook[0]?.order; // may be undefined
+        }),
+      // Fetch app order
+      iexec.orderbook
+        .fetchAppOrderbook(vApp, {
+          minTag: ['tee', 'scone'],
+          maxTag: ['tee', 'scone'],
+          workerpool: vWorkerpool,
+        })
+        .then((appOrderbook) => {
+          const desiredPriceAppOrderbook = appOrderbook.orders.filter(
+            (order) => order.order.appprice <= vAppMaxPrice
+          );
+          const desiredPriceAppOrder = desiredPriceAppOrderbook[0]?.order;
+          if (!desiredPriceAppOrder) {
+            throw new Error('No App order found for the desired price');
+          }
+          return desiredPriceAppOrder;
+        }),
+      // Fetch workerpool order for App or AppWhitelist
+      Promise.all([
+        // for app
+        iexec.orderbook.fetchWorkerpoolOrderbook({
+          workerpool: vWorkerpool,
+          app: vApp,
+          dataset: vProtectedData,
+          requester: requester, // public orders + user specific orders
+          isRequesterStrict: useVoucher, // If voucher, we only want user specific orders
+          minTag: ['tee', 'scone'],
+          maxTag: ['tee', 'scone'],
+          category: 0,
+        }),
+        // for app whitelist
+        iexec.orderbook.fetchWorkerpoolOrderbook({
+          workerpool: vWorkerpool === ethers.ZeroAddress ? 'any' : vWorkerpool,
+          app: vUserWhitelist,
+          dataset: vProtectedData,
+          requester: requester, // public orders + user specific orders
+          isRequesterStrict: useVoucher, // If voucher, we only want user specific orders
+          minTag: ['tee', 'scone'],
+          maxTag: ['tee', 'scone'],
+          category: 0,
+        }),
+      ]).then(
+        ([workerpoolOrderbookForApp, workerpoolOrderbookForAppWhitelist]) => {
+          const desiredPriceWorkerpoolOrder = filterWorkerpoolOrders({
+            workerpoolOrders: [
+              ...workerpoolOrderbookForApp.orders,
+              ...workerpoolOrderbookForAppWhitelist.orders,
+            ],
+            workerpoolMaxPrice: vWorkerpoolMaxPrice,
+            useVoucher: vUseVoucher,
+            userVoucher,
+          });
+          if (!desiredPriceWorkerpoolOrder) {
+            throw new Error('No Workerpool order found for the desired price');
+          }
+          return desiredPriceWorkerpoolOrder;
+        }
+      ),
+    ]);
+
+    if (!workerpoolorder) {
+      throw new Error('No Workerpool order found for the desired price');
+    }
+
+    const datasetorder = datasetorderForApp || datasetorderForWhitelist;
+    if (!datasetorder) {
+      throw new Error('No Dataset order found for the desired price');
+    }
     vOnStatusUpdate({
-      title: 'FETCH_APP_ORDERBOOK',
+      title: 'FETCH_ORDERS',
       isDone: true,
     });
-
-    vOnStatusUpdate({
-      title: 'FETCH_WORKERPOOL_ORDERBOOK',
-      isDone: false,
-    });
-    const workerpoolOrderbook = await iexec.orderbook.fetchWorkerpoolOrderbook({
-      workerpool: vWorkerpool === ethers.ZeroAddress ? 'any' : vWorkerpool, // if address zero was chosen use any workerpool
-      app: vApp,
-      dataset: vProtectedData,
-      minTag: SCONE_TAG,
-      maxTag: SCONE_TAG,
-    });
-    vOnStatusUpdate({
-      title: 'FETCH_WORKERPOOL_ORDERBOOK',
-      isDone: true,
-    });
-
-    const underMaxPriceOrders = fetchOrdersUnderMaxPrice(
-      datasetOrderbook,
-      appOrderbook,
-      workerpoolOrderbook,
-      vMaxPrice
-    );
-
     vOnStatusUpdate({
       title: 'PUSH_REQUESTER_SECRET',
       isDone: false,
@@ -179,13 +269,13 @@ export const processProtectedData = async ({
     });
     const requestorderToSign = await iexec.order.createRequestorder({
       app: vApp,
-      category: underMaxPriceOrders.workerpoolorder.category,
+      category: workerpoolorder.category,
       dataset: vProtectedData,
-      appmaxprice: underMaxPriceOrders.apporder.appprice,
-      datasetmaxprice: underMaxPriceOrders.datasetorder.datasetprice,
-      workerpoolmaxprice: underMaxPriceOrders.workerpoolorder.workerpoolprice,
+      appmaxprice: apporder.appprice,
+      datasetmaxprice: datasetorder.datasetprice,
+      workerpoolmaxprice: workerpoolorder.workerpoolprice,
       tag: SCONE_TAG,
-      workerpool: underMaxPriceOrders.workerpoolorder.workerpool,
+      workerpool: workerpoolorder.workerpool,
       params: {
         iexec_input_files: vInputFiles,
         iexec_secrets: secretsId,
@@ -193,10 +283,22 @@ export const processProtectedData = async ({
       },
     });
     const requestorder = await iexec.order.signRequestorder(requestorderToSign);
-    const { dealid, txHash } = await iexec.order.matchOrders({
+
+    const orders = {
       requestorder,
-      ...underMaxPriceOrders,
-    });
+      workerpoolorder: workerpoolorder,
+      apporder: apporder,
+      datasetorder: datasetorder,
+    };
+    const matchOptions: MatchOptions = {
+      useVoucher: vUseVoucher,
+      ...(vVoucherOwner ? { voucherAddress: userVoucher?.address } : {}),
+    };
+
+    const { dealid, txHash } = await iexec.order.matchOrders(
+      orders,
+      matchOptions
+    );
     const taskId = await iexec.deal.computeTaskId(dealid, 0);
 
     vOnStatusUpdate({