+++ packages/sdk/src/lib/dataProtectorSharing/smartContract/getPocoContract.ts
@@ -1,9 +1,7 @@
 import type { IExecModule } from 'iexec';
-import type { IExecPocoDelegate } from '../../../../generated/typechain/sharing/interfaces/IExecPocoDelegate.js';
+import type { IPoCo } from '../../../../generated/typechain/sharing/interfaces/IPoCo.js';
 
-export async function getPocoContract(
-  iexec: IExecModule
-): Promise<IExecPocoDelegate> {
+export async function getPocoContract(iexec: IExecModule): Promise<IPoCo> {
   const client = await iexec.config.resolveContractsClient();
-  return client.getIExecContract() as unknown as IExecPocoDelegate;
+  return client.getIExecContract() as unknown as IPoCo;
 }