+++ packages/sharing-smart-contract/test/e2e/protectedDataSharing.test.js
@@ -7,12 +7,15 @@ import { deploySCFixture } from './utils/loadFixture.test.js';
 const { ethers } = pkg;
 
 describe('ProtectedDataSharing', () => {
-  describe('AccessControl', () => {
-    it('should set the DEFAULT_ADMIN_ROLE to defaultAdmin', async () => {
-      const { dataProtectorSharingContract, owner, addr1 } = await loadFixture(deploySCFixture);
-      const DEFAULT_ADMIN_ROLE = ethers.toBeHex(0, 32);
-      expect(await dataProtectorSharingContract.hasRole(DEFAULT_ADMIN_ROLE, owner.address)).to.be.true;
-      expect(await dataProtectorSharingContract.hasRole(DEFAULT_ADMIN_ROLE, addr1.address)).to.be.false;
+    describe('AccessControl', () => {
+        it('should set the DEFAULT_ADMIN_ROLE to defaultAdmin', async () => {
+            const { dataProtectorSharingContract, owner, addr1 } =
+                await loadFixture(deploySCFixture);
+            const DEFAULT_ADMIN_ROLE = ethers.toBeHex(0, 32);
+            expect(await dataProtectorSharingContract.hasRole(DEFAULT_ADMIN_ROLE, owner.address)).to
+                .be.true;
+            expect(await dataProtectorSharingContract.hasRole(DEFAULT_ADMIN_ROLE, addr1.address)).to
+                .be.false;
+        });
     });
-  });
 });