+++ packages/sharing-smart-contract/test/e2e/utils/utils.js
@@ -1,8 +1,8 @@
 export const getEventFromLogs = (eventName, logs, { strict = true }) => {
-  const eventFound = logs.find(log => log.eventName === eventName);
-  if (!eventFound) {
-    if (strict) throw new Error(`Unknown event ${eventName}`);
-    return undefined;
-  }
-  return eventFound;
+    const eventFound = logs.find((log) => log.eventName === eventName);
+    if (!eventFound) {
+        if (strict) throw new Error(`Unknown event ${eventName}`);
+        return undefined;
+    }
+    return eventFound;
 };