+++ packages/smart-contract/CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 
 ## Next
 
+### Added
+
+- Support for Arbitrum and Avalanche Fuji testnets (#429).
+
+### Changed
+
+- Integrated Hardhat Ignition for CREATE2 deployment using the CreateX factory for new chain deployments (#429).
+
 ### Changed
 
 - update blockscout url